+++ src/main/java/hu/ppke/itk/nlpg/purepos/model/SuffixGuesser.java
@@ -63,6 +63,6 @@ public abstract class SuffixGuesser<W, T> implements ISuffixGuesser<W, T>,
 		return max.getKey();
 	}
 
-	public void setMapper(IMapper<T> mapper) {
+	public void setTagMapper(ITagMapper<T> mapper) {
 	}
 }